Allow Kagenti UI to use self-signed certs, optionally.
On a fresh openshift install, we may have a self-signed cert. This is
often desirable in testing or development situations, but breaks logins
to the kagenti UI. Here we add a flag that allows kagenti ui to use the
cert bundle generated by the openshift ingress operator, or other
self-signed certs.